docs: translate all JavaDocs in API module * refactor: Enchantment compatibilities & add rarity * fix: Enchantments rarity * refactor: Do not use short on enchantments * refactor: API module

+++ Allay-API/src/main/java/org/allaymc/api/item/enchantment/EnchantmentInstance.java
@@ -11,16 +11,16 @@ public interface EnchantmentInstance {
 
     EnchantmentType getType();
 
-    short getLevel();
+    int getLevel();
 
-    default EnchantmentInstance setLevel(short level) {
+    default EnchantmentInstance setLevel(int level) {
         return getType().createInstance(level);
     }
 
     default NbtMap saveNBT() {
         return NbtMap.builder()
-                .putShort("id", getType().getId())
-                .putShort("lvl", getLevel())
+                .putShort("id", (short) getType().getId())
+                .putShort("lvl", (short) getLevel())
                 .build();
     }
 }
